implemented AutoUpdatableAdapter

+++ app/src/main/java/com/kvest/diffutils_with_coroutines/Item.kt
@@ -3,4 +3,4 @@ package com.kvest.diffutils_with_coroutines
 /**
  * Created by kvest on 7/11/2017.
  */
-data class Item(val id: Long, val name: String)
+data class Item(val id: Int, val name: String)
